+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+* Incremental writing support (i.e. appending a single revision with all the
+  changes to an existing document) to [HexaPDF::Writer] and [HexaPDF::Document]
 * CLI command `hexapdf split` to split a PDF file into individual pages
 * [HexaPDF::Revisions#parser] for accessing the parser object that is created
   when a document is read from an IO stream